Implement operators for SparseVector
